+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+### unreleased
+### Changed
+* Changed supported Blender version to 4.0 or later.
+
 ### 0.2.0 - 2023-08-10
 ### Added
 * Added `models/colors.ldr` for testing all current LDraw colors.